HomepageScreen, LoginScreen in progress.

+++ FocusLit-Android/app/src/main/java/com/focus/lit/data/model/UserAnalytics.kt
@@ -3,11 +3,11 @@ package com.focus.lit.data.model
 data class UserAnalytics(
     val  id: Int,
 
-    val  score: Int,
+    val  score: Double,
 
     val  streak: Int,
 
-    val  totalWorkDuration: Int,
+    val  totalWorkDuration: Double,
 
     val  userRank: Int
 )
